BclconvertReports R6 Class

Reads and writes tidy versions of files within the Reports directory output from BCLConvert v4.2.7. See the DRAGEN v4.2 documentation at https://support-docs.illumina.com/SW/dragen_v42/Content/SW/DRAGEN/OutputFiles.htm.

Public fields

path

Path to the Reports directory.

contents

Tibble with file path, basename, and size.

Methods

Public methods

BclconvertReports$new()

BclconvertReports$print()

BclconvertReports$read_adaptercyclemetrics()

BclconvertReports$read_adaptermetrics()

BclconvertReports$read_demultiplexstats()

BclconvertReports$read_demultiplextilestats()

BclconvertReports$read_qualitymetrics()

BclconvertReports$read_qualitytilemetrics()

BclconvertReports$read_indexhoppingcounts()

BclconvertReports$read_topunknownbarcodes()

BclconvertReports$read_fastqlist()

BclconvertReports$read()

BclconvertReports$write()

BclconvertReports$clone()

Method new()

Create a new BclconvertReports object.

Usage

BclconvertReports$new(path = NULL)

Arguments

path

Path to the Reports directory.

Method print()

Print details about the BclconvertReports directory.

Usage

BclconvertReports$print(...)

Arguments

...

(ignored).

Method read_adaptercyclemetrics()

Read Adapter_Cycle_Metrics.csv file.

lane: lane number.

sampleid: sample ID from sample sheet.

indexes: index/index2 from sample sheet for this sample.

read: read number.

cycle: cycle number.

cluster_n: number of clusters where the adapter was detected to begin precisely at this cycle.

cluster_pct: percentage of all clusters where the adapter was detected to begin precisely at this cycle.

Usage

BclconvertReports$read_adaptercyclemetrics(x)

Arguments

x

(character(1))
Path to Adapter_Cycle_Metrics.csv file.

Method read_adaptermetrics()

Read Adapter_Metrics.csv file.

lane: lane number.

sampleid: sample ID from sample sheet.

indexes: index/index2 from sample sheet for this sample.

readnum: read number.

adapter_bases: total number of bases trimmed as adapter from the read.

sample_bases: total number of bases not trimmed from the read.

adapter_bases_pct: percentage of bases trimmed as adapter from the read.

Usage

BclconvertReports$read_adaptermetrics(x)

Arguments

x

(character(1))
Path to Adapter_Metrics.csv file.

Method read_demultiplexstats()

Read Demultiplex_Stats.csv file.

lane: lane number.

sampleid: sample ID from sample sheet.

indexes: index/index2 from sample sheet for this sample.

reads_n: total number of pass-filter reads mapping to this sample for the lane.

perfect_idxreads_n: number of mapped reads with barcodes matching the indexes exactly.

one_mismatch_idxreads_n: number of mapped reads with barcodes matched with one base mismatched.

two_mismatch_idxreads_n: number of mapped reads with barcodes matched with two bases mismatched.

reads_pct: percentage of pass-filter reads mapping to this sample for the lane.

perfect_idxreads_pct: percentage of mapped reads with barcodes matching the indexess exactly.

one_mismatch_idxreads_pct: percentage of mapped reads with one mismatch to the indexes.

two_mismatch_idxreads_pct: percentage of mapped reads with two mismatches to the indexes.

Usage

BclconvertReports$read_demultiplexstats(x)

Arguments

x

(character(1))
Path to Demultiplex_Stats.csv file.

Method read_demultiplextilestats()

Read Demultiplex_Tile_Stats.csv file.

lane: lane number.

sampleid: sample ID from sample sheet.

indexes: index/index2 from sample sheet for this sample.

tile: tile number.

reads_n: total number of pass-filter reads mapping to this sample for the lane.

reads_pct: percentage of pass-filter reads mapping to this sample for the lane.

Usage

BclconvertReports$read_demultiplextilestats(x)

Arguments

x

(character(1))
Path to Demultiplex_Tile_Stats.csv file.

Method read_qualitymetrics()

Read Quality_Metrics.csv file.

lane: lane number.

sampleid: sample ID from sample sheet.

indexes: index/index2 from sample sheet for this sample.

readnum: read number (1 or 2).

yield: number of bases mapping.

yieldq30: number of bases with quality score >= 30 mapping.

qscore_sum: sum of quality scores of bases mapping.

qscore_mean_pf: mean quality score of bases mapping.

q30_pct: percentage of bases with quality score >= 30 mapping.

Usage

BclconvertReports$read_qualitymetrics(x)

Arguments

x

(character(1))
Path to Quality_Metrics.csv file.

Method read_qualitytilemetrics()

Read Quality_Tile_Metrics.csv file.

lane: lane number.

sampleid: sample ID from sample sheet.

indexes: index/index2 from sample sheet for this sample.

readnum: read number (1 or 2).

tile: tile number.

yield: number of bases mapping.

yieldq30: number of bases with quality score >= 30 mapping.

qscore_sum: sum of quality scores of bases mapping.

qscore_mean_pf: mean quality score of bases mapping.

q30_pct: percentage of bases with quality score >= 30 mapping.

Usage

BclconvertReports$read_qualitytilemetrics(x)

Arguments

x

(character(1))
Path to Quality_Tile_Metrics.csv file.

Method read_indexhoppingcounts()

Read Index_Hopping_Counts.csv file.

lane: lane number.

sampleid: sample ID from sample sheet.

indexes: index/index2 from sample sheet for this sample.

reads_n: total number of pass-filter reads mapping to the indexes.

reads_hopped_pct: percentage of hopped pass-filter reads mapping to the indexes.

reads_pct: percentage of all pass-filter reads mapping to the indexes.

Usage

BclconvertReports$read_indexhoppingcounts(x)

Arguments

x

(character(1))
Path to Index_Hopping_Counts.csv file.

Method read_topunknownbarcodes()

Read Top_Unknown_Barcodes.csv file.

lane: lane number.

indexes: index/index2 of this unlisted sequence.

reads_n: total number of pass-filter reads mapping to the indexes.

unknownbcodes_pct: percentage of unknown pass-filter reads mapping to the indexes.

Usage

BclconvertReports$read_topunknownbarcodes(x)

Arguments

x

(character(1))
Path to Top_Unknown_Barcodes.csv file.

Method read_fastqlist()

Read fastq_list.csv file.

rgid: read group.

rgsm: sample ID.

rglb: library.

lane: flow cell lane.

readnum: read number (1 or 2).

filepath: path to the FASTQ file.

Usage

BclconvertReports$read_fastqlist(x)

Arguments

x

(character(1))
Path to fastq_list.csv file.

Method read()

Reads contents of Reports directory output by BCLConvert.

Usage

BclconvertReports$read()

Returns

A list of tibbles.

Method write()

Writes tidied contents of Reports directory output by BCLConvert.

Usage

BclconvertReports$write(
  d,
  out_dir = NULL,
  prefix,
  out_format = "tsv",
  drid = NULL
)

Arguments

d

Parsed object from self$read().

out_dir

Output directory.

prefix

Prefix of output file(s).

out_format

Format of output file(s).

drid

dracarys ID to use for the dataset (e.g. wfrid.123, prid.456).

Method clone()

The objects of this class are cloneable with this method.

Usage

BclconvertReports$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) { # \dontrun{
p1 <- "240816_A01052_0220_AHM7VHDSXC/202408195d4f2fc4/Reports"
b <- here::here("nogit/bcl_convert", p1) |>
  BclconvertReports$new()
b$path
b$contents
d <- b$read()
b$write(d, out_dir = tempdir(), prefix = "sampleA", out_format = "tsv")
} # }