CoverageMetricsFile R6 Class

Contains methods for reading and displaying contents of the wgs_coverage_metrics_<phenotype>.csv file output from DRAGEN. This file contains read depth of coverage metrics.

Super class

dracarys::File -> CoverageMetricsFile

Methods

Public methods

CoverageMetricsFile$read()

CoverageMetricsFile$clone()

Inherited methods

dracarys::File$bname()

dracarys::File$initialize()

dracarys::File$print()

dracarys::File$type()

Method read()

Reads the wgs_coverage_metrics_<phenotype>.csv file output from DRAGEN.

Usage

CoverageMetricsFile$read()

Returns

tibble with the following columns:

label: file label.

var: variable name.

var_abbrev: variable abbreviation.

pct: percentage value.

count: count value.

Method clone()

The objects of this class are cloneable with this method.

Usage

CoverageMetricsFile$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

x1 <- system.file("extdata/wgs/SEQC-II.wgs_coverage_metrics_normal.csv.gz", package = "dracarys")
x2 <- system.file("extdata/wgs/SEQC-II.wgs_coverage_metrics_tumor.csv.gz", package = "dracarys")
cm1 <- CoverageMetricsFile$new(x1)
cm2 <- CoverageMetricsFile$new(x2)
read(cm1)
#> # A tibble: 27 × 5
#>    label     var                                            var_a…¹    pct count
#>    <chr>     <chr>                                          <chr>    <dbl> <dbl>
#>  1 SEQC-II_N Uniformity of coverage (PCT > 0.2*mean) over … unifor…  95.8     NA
#>  2 SEQC-II_N PCT of genome with coverage [100x: inf)        %genom…   0.25    NA
#>  3 SEQC-II_N PCT of genome with coverage [ 50x: inf)        %genom…  43.8     NA
#>  4 SEQC-II_N PCT of genome with coverage [ 20x: inf)        %genom…  93.2     NA
#>  5 SEQC-II_N PCT of genome with coverage [ 15x: inf)        %genom…  95.0     NA
#>  6 SEQC-II_N PCT of genome with coverage [ 10x: inf)        %genom…  95.8     NA
#>  7 SEQC-II_N PCT of genome with coverage [  3x: inf)        %genom…  96.7     NA
#>  8 SEQC-II_N PCT of genome with coverage [  1x: inf)        %genom…  97.2     NA
#>  9 SEQC-II_N PCT of genome with coverage [  0x: inf)        %genom… 100       NA
#> 10 SEQC-II_N PCT of genome with coverage [ 50x:100x)        %genom…  43.6     NA
#> # … with 17 more rows, and abbreviated variable name ¹​var_abbrev
read(cm2)
#> # A tibble: 27 × 5
#>    label     var                                             var_a…¹   pct count
#>    <chr>     <chr>                                           <chr>   <dbl> <dbl>
#>  1 SEQC-II_T Uniformity of coverage (PCT > 0.2*mean) over g… unifor…  96.1    NA
#>  2 SEQC-II_T PCT of genome with coverage [100x: inf)         %genom…  22.3    NA
#>  3 SEQC-II_T PCT of genome with coverage [ 50x: inf)         %genom…  89.5    NA
#>  4 SEQC-II_T PCT of genome with coverage [ 20x: inf)         %genom…  96.0    NA
#>  5 SEQC-II_T PCT of genome with coverage [ 15x: inf)         %genom…  96.2    NA
#>  6 SEQC-II_T PCT of genome with coverage [ 10x: inf)         %genom…  96.5    NA
#>  7 SEQC-II_T PCT of genome with coverage [  3x: inf)         %genom…  97.1    NA
#>  8 SEQC-II_T PCT of genome with coverage [  1x: inf)         %genom…  97.5    NA
#>  9 SEQC-II_T PCT of genome with coverage [  0x: inf)         %genom… 100      NA
#> 10 SEQC-II_T PCT of genome with coverage [ 50x:100x)         %genom…  67.3    NA
#> # … with 17 more rows, and abbreviated variable name ¹​var_abbrev