MappingMetricsFile R6 Class

Contains methods for reading and displaying contents of the mapping_metrics.csv file output from DRAGEN. This file contains mapping and aligning metrics, like the metrics computed by the Samtools Flagstat command. These metrics are available on an aggregate level (over all input data), and on a per read group level. Unless explicitly stated, the metrics units are in reads (i.e., not in terms of pairs or alignments).

Super class

dracarys::File -> MappingMetricsFile

Methods

Public methods

MappingMetricsFile$read()

MappingMetricsFile$clone()

Inherited methods

dracarys::File$bname()

dracarys::File$initialize()

dracarys::File$print()

dracarys::File$type()

Method read()

Reads the mapping_metrics.csv file output from DRAGEN.

Usage

MappingMetricsFile$read()

Returns

tibble with the following columns: - category: summary or read group - Phenotype: e.g. tumor, normal - RG: read group - var: metric variable - var_abbrev: metric variable abbreviation - count: count of reads - pct: percentage of reads

Method clone()

The objects of this class are cloneable with this method.

Usage

MappingMetricsFile$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

x <- system.file("extdata/wgs/SEQC-II.mapping_metrics.csv.gz", package = "dracarys")
mm <- MappingMetricsFile$new(x)
mm$read() # or read(mm)
#> # A tibble: 218 × 7
#>    category Phenotype RG    var                         var_abbrev   count   pct
#>    <chr>    <chr>     <chr> <chr>                       <chr>        <dbl> <dbl>
#>  1 summary  tumor     TOTAL Total Reads per RG          Tot         2.20e9 100  
#>  2 summary  tumor     TOTAL Number of duplicate marked… Dup         4.43e8  20.2
#>  3 summary  tumor     TOTAL Number of duplicate marked… Dup Rem    NA       NA  
#>  4 summary  tumor     TOTAL Number of unique reads (ex… Unique      1.76e9  79.8
#>  5 summary  tumor     TOTAL Reads with mate sequenced   Mated       2.20e9 100  
#>  6 summary  tumor     TOTAL Reads without mate sequenc… noMated     0        0  
#>  7 summary  tumor     TOTAL QC-failed reads             Failed      0        0  
#>  8 summary  tumor     TOTAL Mapped reads                Mapped      2.20e9  99.8
#>  9 summary  tumor     TOTAL Mapped reads adjusted for … Mapped re…  2.20e9  99.8
#> 10 summary  tumor     TOTAL Mapped reads R1             R1map       1.10e9 100. 
#> # ℹ 208 more rows