ReplayFile R6 Class

Contains methods for reading contents of the replay.json file output from DRAGEN, which contains the DRAGEN command line, parameters and version for the specific run.

Super class

dracarys::File -> ReplayFile

Methods

Public methods

ReplayFile$read()

ReplayFile$clone()

Inherited methods

dracarys::File$bname()

dracarys::File$initialize()

dracarys::File$print()

dracarys::File$type()

Method read()

Reads the replay.json file.

Usage

ReplayFile$read()

Returns

A list with the following elements:

command_line: character of DRAGEN command line used.

dragen_config: tibble of parameters used for the DRAGEN run.

system: tibble with dragen_version, nodename, and kernel_release.

label: character of sample label (inferred from file name)

hash_table_build: tibble with details about the DRAGEN hash table build.

Method clone()

The objects of this class are cloneable with this method.

Usage

ReplayFile$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

x <- system.file("extdata/wgs/SEQC-II-replay.json.gz", package = "dracarys")
r <- ReplayFile$new(x)
r$read() # or read(r)
#> $command_line
#> [1] "/opt/edico/bin/dragen --enable-variant-caller true --intermediate-results-dir /ephemeral/intermediate-results/ --cnv-use-somatic-vc-baf true --enable-cnv true --enable-duplicate-marking true --enable-map-align-output true --enable-sv true --fastq-list /ZnDLLh/fastq_list.csv --lic-instance-id-location /opt/instance-identity --output-directory SEQC-II_dragen_somatic --output-file-prefix SEQC-II --ref-dir /ephemeral/ref/hg38-v8-altaware-cnv-anchored/ --tumor-fastq-list /ZnDLLh/tumor_fastq_list.csv"
#> 
#> $hash_table_build
#> # A tibble: 3 × 2
#>   name                 value                                                    
#>   <chr>                <chr>                                                    
#> 1 sw_version           01.003.044.3.5.3-38-gefdaeaff                            
#> 2 hash_table_version   8                                                        
#> 3 command_line_options /opt/edico/bin/dragen --lic-instance-id-location /root/.…
#> 
#> $dragen_config
#> # A tibble: 470 × 2
#>    name                    value
#>    <chr>                   <chr>
#>  1 Aligner.align-direction 4    
#>  2 Aligner.aln-en-mask     15   
#>  3 Aligner.aln-enable      1    
#>  4 Aligner.backtrace-delay 8    
#>  5 Aligner.bed-score-bonus 0    
#>  6 Aligner.chicken         0    
#>  7 Aligner.dedup-min-qual  15   
#>  8 Aligner.disable-lfsr    0    
#>  9 Aligner.en-alt-hap-aln  1    
#> 10 Aligner.en-chimeric-aln 1    
#> # … with 460 more rows
#> 
#> $system
#> # A tibble: 3 × 2
#>   name           value                                     
#>   <chr>          <chr>                                     
#> 1 dragen_version 05.021.609.3.9.3                          
#> 2 nodename       trn.6f62637e545a418f8dfe9ce587e571d7-gt4mv
#> 3 kernel_release 5.4.141-67.229.amzn2.x86_64               
#> 
#> $label
#> [1] "SEQC-II"
#>