TimeMetricsFile R6 Class

Contains methods for reading contents of the time_metrics.csv file output from DRAGEN, which contains a breakdown of the run duration for each DRAGEN process.

Super class

dracarys::File -> TimeMetricsFile

Methods

Public methods

TimeMetricsFile$read()

TimeMetricsFile$clone()

Inherited methods

dracarys::File$bname()

dracarys::File$initialize()

dracarys::File$print()

dracarys::File$type()

Method read()

Reads the time_metrics.csv file.

Usage

TimeMetricsFile$read()

Returns

tibble with the following columns:

Step: DRAGEN step

Time: time in HH:MM

Method clone()

The objects of this class are cloneable with this method.

Usage

TimeMetricsFile$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

x <- system.file("extdata/wgs/SEQC-II.time_metrics.csv.gz", package = "dracarys")
tm <- TimeMetricsFile$new(x)
tm$read() # or read(tm)
#> # A tibble: 19 × 3
#>    Label   Step                           Time 
#>    <chr>   <chr>                          <chr>
#>  1 SEQC-II Loading Reference              00:01
#>  2 SEQC-II Aligning Reads                 00:35
#>  3 SEQC-II Duplicate Marking              00:01
#>  4 SEQC-II Sorting and Marking Duplicates 01:31
#>  5 SEQC-II DRAGStr Calibration            00:00
#>  6 SEQC-II Saving Map/Align Output        01:31
#>  7 SEQC-II Partial Reconfiguration        00:00
#>  8 SEQC-II Variant Calling                01:31
#>  9 SEQC-II Calculating Target Counts      00:36
#> 10 SEQC-II Correcting GC Bias             00:00
#> 11 SEQC-II Normalizing Case Sample        00:00
#> 12 SEQC-II Performing Segmentation        00:00
#> 13 SEQC-II Generating Somatic Model Input 00:17
#> 14 SEQC-II Fitting Somatic Model          00:00
#> 15 SEQC-II Generating CNV Calls           00:00
#> 16 SEQC-II Generating CNV Track Files     00:00
#> 17 SEQC-II Partitioning                   00:35
#> 18 SEQC-II Structural Variant Calling     01:38
#> 19 SEQC-II Total Runtime                  04:10