TsoCombinedVariantOutputFile R6 Class

Contains methods for reading and displaying contents of the CombinedVariantOutput.tsv file output from TSO.

Super class

dracarys::File -> TsoCombinedVariantOutputFile

Methods

Public methods

TsoCombinedVariantOutputFile$read()

TsoCombinedVariantOutputFile$clone()

Inherited methods

dracarys::File$bname()

dracarys::File$initialize()

dracarys::File$print()

dracarys::File$type()

Method read()

Reads the CombinedVariantOutput.tsv file output from TSO.

Usage

TsoCombinedVariantOutputFile$read()

Returns

list of following tibbles:

FragmentLength

Count

Method clone()

The objects of this class are cloneable with this method.

Usage

TsoCombinedVariantOutputFile$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

x <- system.file("extdata/tso/sample705_CombinedVariantOutput.tsv", package = "dracarys")
d <- TsoCombinedVariantOutputFile$new(x)
d$read() # or read(d)
#> $`Analysis Details`
#> # A tibble: 4 × 2
#>   variable         value                   
#>   <chr>            <chr>                   
#> 1 DNA Sample ID    PTC_ctTSO220118_L2200054
#> 2 Output Date      2022-02-19              
#> 3 Output Time      21:51:38                
#> 4 Pipeline Version 0.0.0.1                 
#> 
#> $`Sequencing Run Details`
#> # A tibble: 0 × 0
#> 
#> $TMB
#> # A tibble: 3 × 2
#>   variable                            value
#>   <chr>                               <chr>
#> 1 Total TMB                           143.8
#> 2 Coding Region Size in Megabases     1.22 
#> 3 Number of Passing Eligible Variants 176  
#> 
#> $MSI
#> # A tibble: 1 × 2
#>   variable value
#>   <chr>    <chr>
#> 1 SUM_JSD  1.401
#> 
#> $`Copy Number Variants`
#> # A tibble: 0 × 0
#> 
#> $`DNA Fusions`
#> # A tibble: 2 × 7
#>   `Gene 1` `Gene 2` `Breakend 1`   `Breakend 2`   Supporting R…¹ Total…² Varia…³
#>   <chr>    <chr>    <chr>          <chr>          <chr>          <chr>   <chr>  
#> 1 ALK      TPR      chr2:29446944  chr1:186325172 25             11760   0.0021…
#> 2 RET      NCOA4    chr10:43610592 chr10:51584250 21             8279    0.0025…
#> # … with abbreviated variable names ¹​`Supporting Read Pairs`,
#> #   ²​`Total Read Pairs`, ³​`Variant Allele Frequency`
#> 
#> $`Small Variants`
#> # A tibble: 1,980 × 11
#>    Gene    Chrom…¹ Genom…² Refer…³ Alter…⁴ Allel…⁵ Depth P-Dot…⁶ C-Dot…⁷ Conse…⁸
#>    <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr> <chr>   <chr>   <chr>  
#>  1 TNFRSF… chr1    2488153 A       G       0.4885  4917  NP_003… NM_003… missen…
#>  2 NA      chr1    4367323 G       A       0.9998  5272  NA      NA      NA     
#>  3 PIK3CD  chr1    9778888 G       A       0.0016  7923  NP_005… NM_005… missen…
#>  4 PIK3CD  chr1    9781605 G       A       0.0017  7471  NP_005… NM_005… missen…
#>  5 PIK3CD  chr1    9781637 G       A       0.0013  6136  NP_005… NM_005… stop_g…
#>  6 PIK3CD  chr1    9784354 C       T       0.0019  4699  NM_005… NM_005… synony…
#>  7 MTOR    chr1    112050… C       T       0.9997  5856  NM_004… NM_004… synony…
#>  8 MTOR    chr1    112762… C       G       0.0032  950   NP_004… NM_004… missen…
#>  9 MTOR    chr1    112887… G       A       0.9998  8997  NM_004… NM_004… synony…
#> 10 MTOR    chr1    112925… C       T       0.0021  3809  NP_004… NM_004… missen…
#> # … with 1,970 more rows, 1 more variable: `Affected Exon(s)` <chr>, and
#> #   abbreviated variable names ¹​Chromosome, ²​`Genomic Position`,
#> #   ³​`Reference Call`, ⁴​`Alternative Call`, ⁵​`Allele Frequency`,
#> #   ⁶​`P-Dot Notation`, ⁷​`C-Dot Notation`, ⁸​`Consequence(s)`
#>