TsoSampleAnalysisResultsFile R6 Class

Contains methods for reading and displaying contents of the SampleAnalysisResults.json.gz file output from TSO.

Super class

dracarys::File -> TsoSampleAnalysisResultsFile

Methods

Public methods

TsoSampleAnalysisResultsFile$read()

TsoSampleAnalysisResultsFile$write()

TsoSampleAnalysisResultsFile$clone()

Inherited methods

dracarys::File$bname()

dracarys::File$initialize()

dracarys::File$print()

dracarys::File$type()

Method read()

Reads the SampleAnalysisResults.json.gz file output from TSO.

Usage

TsoSampleAnalysisResultsFile$read()

Returns

list of tibbles

Method write()

Writes a tidy version of the SampleAnalysisResults.json.gz file output from TSO.

Usage

TsoSampleAnalysisResultsFile$write(d, out_dir, prefix, out_format = "tsv")

Arguments

d

Parsed object from self$read().

out_dir

Output directory.

prefix

Prefix of output file(s).

out_format

Format of output file(s) (one of 'tsv' (def.), 'parquet', 'both').

Method clone()

The objects of this class are cloneable with this method.

Usage

TsoSampleAnalysisResultsFile$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

x <- system.file("extdata/tso/sample705_SampleAnalysisResults.json.gz", package = "dracarys")
res <- TsoSampleAnalysisResultsFile$new(x)
d_parsed <- res$read() # or read(res)
res$write(d_parsed, tempfile(), "both")
#> $sample_info
#> # A tibble: 1 × 4
#>   sampleId                 analysisDate analysisTime analysisName              
#>   <chr>                    <chr>        <chr>        <chr>                     
#> 1 PTC_ctTSO220118_L2200054 2022-02-19   21:49:55     ctTSO220118-220207_18Feb22
#> 
#> $qc
#> # A tibble: 1 × 39
#>   dna_qc_CONTAMINATION_SCORE_value dna_qc_CONTAMINATION…¹ dna_qc_CONTAMINATION…²
#>                              <dbl>                  <dbl>                  <dbl>
#> 1                              664                      0                   1672
#> # ℹ abbreviated names: ¹​dna_qc_CONTAMINATION_SCORE_LSL,
#> #   ²​dna_qc_CONTAMINATION_SCORE_USL
#> # ℹ 36 more variables: dna_qc_CONTAMINATION_P_VALUE_value <dbl>,
#> #   dna_qc_CONTAMINATION_P_VALUE_LSL <dbl>,
#> #   dna_qc_CONTAMINATION_P_VALUE_USL <dbl>,
#> #   dna_qc_smallv_tmb_MEDIAN_EXON_COVERAGE_value <dbl>,
#> #   dna_qc_smallv_tmb_MEDIAN_EXON_COVERAGE_LSL <dbl>, …
#> 
#> $biomarkers
#> # A tibble: 1 × 5
#>   msi_pct_unstable_sites msi_SumJsd tmb_per_mb tmb_coding_region_sizemb
#>                    <dbl>      <dbl>      <dbl>                    <dbl>
#> 1                    2.2       1.40       144.                     1.22
#> # ℹ 1 more variable: tmb_somatic_coding_variants_count <dbl>
#> 
#> $sw_conf_datasources
#> # A tibble: 7 × 4
#>   name                 version     description                       releaseDate
#>   <chr>                <chr>       <chr>                             <chr>      
#> 1 VEP                  91          BothRefSeqAndEnsembl              2018-03-05 
#> 2 1000 Genomes Project Phase 3 v5a A public catalogue of human vari… 2013-05-27 
#> 3 ClinVar              20190204    A freely accessible, public arch… 2019-02-04 
#> 4 COSMIC               84          somatic mutation and related det… 2018-02-13 
#> 5 dbSNP                151         Identifiers for observed variants 2018-04-23 
#> 6 gnomAD               2.1         Whole genome allele frequencies … 2018-10-17 
#> 7 gnomAD_exome         2.1         Exome allele frequencies from Ge… 2018-10-17 
#> 
#> $sw_conf_other
#> # A tibble: 1 × 7
#>   analysisSoftwareName analysisSoftwareVersion genomeBuild       instanceName
#>   <chr>                <chr>                   <chr>             <chr>       
#> 1 TSO500_ctDNA_RnD     0.0.0.1                 hg19_decoy/grch37 CDxNirvana  
#> # ℹ 3 more variables: nirvanaSoftwareVersion <chr>, genomeAssembly <chr>,
#> #   refseqVersion <chr>
#> 
#> $snv
#> # A tibble: 1,980 × 29
#>    chrom    pos ref   alt       af  qual dp_tot dp_alt transcript source bioType
#>    <chr>  <int> <chr> <chr>  <dbl> <dbl>  <int>  <int> <chr>      <chr>  <chr>  
#>  1 chr1  2.49e6 A     G     0.488    100   4917   2402 NM_003820… RefSeq protei…
#>  2 chr1  4.37e6 G     A     1.00     100   5272   5271 NA         NA     NA     
#>  3 chr1  9.78e6 G     A     0.0016   100   7923     13 NM_005026… RefSeq protei…
#>  4 chr1  9.78e6 G     A     0.0017   100   7471     13 NM_005026… RefSeq protei…
#>  5 chr1  9.78e6 G     A     0.0013   100   6136      8 NM_005026… RefSeq protei…
#>  6 chr1  9.78e6 C     T     0.0019   100   4699      9 NM_005026… RefSeq protei…
#>  7 chr1  1.12e7 C     T     1.00     100   5856   5854 NM_004958… RefSeq protei…
#>  8 chr1  1.13e7 C     G     0.0032   100    950      3 NM_004958… RefSeq protei…
#>  9 chr1  1.13e7 G     A     1.00     100   8997   8995 NM_004958… RefSeq protei…
#> 10 chr1  1.13e7 C     T     0.0021   100   3809      8 NM_004958… RefSeq protei…
#> # ℹ 1,970 more rows
#> # ℹ 18 more variables: aminoAcids <chr>, cdnaPos <chr>, codons <chr>,
#> #   cdsPos <chr>, exons <chr>, geneId <chr>, hgnc <chr>, hgvsc <chr>,
#> #   hgvsp <chr>, isCanonical <lgl>, polyPhenScore <dbl>,
#> #   polyPhenPrediction <chr>, proteinId <chr>, proteinPos <chr>,
#> #   siftScore <dbl>, siftPrediction <chr>, consequence <chr>, introns <chr>
#> 
#> $cnv
#> # A tibble: 1 × 7
#>   foldChange  qual copyNumberType gene  chromosome startPosition endPosition
#>        <dbl> <dbl> <chr>          <chr> <chr>              <int>       <int>
#> 1       0.86    57 <DEL>          ATM   chr11          108098348   108236237
#>