TsoTargetRegionCoverageFile R6 Class

Contains methods for reading and displaying contents of the TargetRegionCoverage.json.gz file output from TSO.

Super class

dracarys::File -> TsoTargetRegionCoverageFile

Methods

Public methods

TsoTargetRegionCoverageFile$read()

TsoTargetRegionCoverageFile$plot()

TsoTargetRegionCoverageFile$clone()

Inherited methods

dracarys::File$bname()

dracarys::File$initialize()

dracarys::File$print()

dracarys::File$type()

Method read()

Reads the TargetRegionCoverage.json.gz file output from TSO.

Usage

TsoTargetRegionCoverageFile$read()

Returns

tibble with the following columns:

ConsensusReadDepth

BasePair

Percentage

Method plot()

Plots the TargetRegionCoverage.json.gz file.

Usage

TsoTargetRegionCoverageFile$plot(min_pct = 2)

Arguments

min_pct

Minimum percentage to be plotted (Default: 2).

Returns

A ggplot2 plot containing read depth on X axis and percentage covered on Y axis.

Method clone()

The objects of this class are cloneable with this method.

Usage

TsoTargetRegionCoverageFile$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

x <- system.file("extdata/tso/sample705.TargetRegionCoverage.json.gz", package = "dracarys")
trc <- TsoTargetRegionCoverageFile$new(x)
trc$read() # or read(trc)
#> # A tibble: 14 × 3
#>    ConsensusReadDepth BasePair Percentage
#>    <chr>              <chr>    <chr>     
#>  1 TargetRegion       1970825  100%      
#>  2 100X               1953719  99.13     
#>  3 250X               1940233  98.45     
#>  4 500X               1922662  97.56     
#>  5 750X               1907303  96.78     
#>  6 1000X              1891110  95.96     
#>  7 1500X              1848614  93.80     
#>  8 2000X              1795480  91.10     
#>  9 2500X              1729811  87.77     
#> 10 3000X              1651245  83.78     
#> 11 4000X              1445949  73.37     
#> 12 5000X              1209504  61.37     
#> 13 8000X              522143   26.49     
#> 14 10000X             192278   9.76      
trc$plot(90) # or plot(trc, 90)
#> Error in loadNamespace(x): there is no package called ‘ggrepel’