TsoTmbTraceTsvFile R6 Class

Contains methods for reading and displaying contents of the TMB_Trace.tsv file output from TSO.

Super class

dracarys::File -> TsoTmbTraceTsvFile

Methods

Public methods

TsoTmbTraceTsvFile$read()

TsoTmbTraceTsvFile$write()

TsoTmbTraceTsvFile$clone()

Inherited methods

dracarys::File$bname()

dracarys::File$initialize()

dracarys::File$print()

dracarys::File$type()

Method read()

Reads the TMB_Trace.tsv file output from TSO.

Usage

TsoTmbTraceTsvFile$read()

Returns

tibble with variants.

Method write()

Writes a tidy version of the TMB_Trace.tsv file output from TSO.

Usage

TsoTmbTraceTsvFile$write(d, out_dir, prefix, out_format = "tsv")

Arguments

d

Parsed object from self$read().

out_dir

Output directory.

prefix

Prefix of output file(s).

out_format

Format of output file(s) (one of 'tsv' (def.), 'parquet', 'both').

Method clone()

The objects of this class are cloneable with this method.

Usage

TsoTmbTraceTsvFile$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

x <- system.file("extdata/tso/sample705_TMB_Trace.tsv", package = "dracarys")
d <- TsoTmbTraceTsvFile$new(x)
d_parsed <- d$read() # or read(d)
d$write(d_parsed, out_dir = tempdir(), prefix = "sample705", out_format = "both")
#> # A tibble: 2,186 × 20
#>    Chromo…¹ Posit…² RefCall AltCall    VAF Depth CytoB…³ GeneN…⁴ Varia…⁵ Cosmi…⁶
#>    <chr>      <int> <chr>   <chr>    <dbl> <dbl> <chr>   <chr>   <chr>   <chr>  
#>  1 chr1      2.49e6 A       G       0.488   4917 1p36.32 LOC100… SNV     COSM41…
#>  2 chr1      4.37e6 G       A       1.00    5272 1p36.32 NA      SNV     NA     
#>  3 chr1      9.78e6 G       A       0.0016  7923 1p36.22 PIK3CD  SNV     NA     
#>  4 chr1      9.78e6 G       A       0.0017  7471 1p36.22 PIK3CD  SNV     NA     
#>  5 chr1      9.78e6 G       A       0.0013  6136 1p36.22 PIK3CD  SNV     NA     
#>  6 chr1      9.78e6 C       T       0.0019  4699 1p36.22 CLSTN1… SNV     NA     
#>  7 chr1      1.12e7 C       T       1.00    5856 1p36.22 MTOR:M… SNV     COSM41…
#>  8 chr1      1.13e7 C       G       0.0032   950 1p36.22 MTOR    SNV     NA     
#>  9 chr1      1.13e7 G       A       1.00    8997 1p36.22 MTOR    SNV     COSM41…
#> 10 chr1      1.13e7 C       T       0.0021  3809 1p36.22 MTOR    SNV     COSM15…
#> # … with 2,176 more rows, 10 more variables: MaxCosmicCount <dbl>,
#> #   AlleleCountsGnomadExome <dbl>, AlleleCountsGnomadGenome <dbl>,
#> #   AlleleCounts1000Genomes <dbl>, MaxDatabaseAlleleCounts <dbl>,
#> #   GermlineFilterDatabase <lgl>, GermlineFilterProxi <lgl>,
#> #   CodingVariant <lgl>, Nonsynonymous <lgl>, IncludedInTMBNumerator <lgl>, and
#> #   abbreviated variable names ¹​Chromosome, ²​Position, ³​CytoBand, ⁴​GeneName,
#> #   ⁵​VariantType, ⁶​CosmicIDs