UmHrdetectTsvFile R6 Class

Contains methods for reading and displaying contents of the hrdetect.tsv.gz file output from umccrise.

Super class

dracarys::File -> UmHrdetectTsvFile

Methods

Public methods

UmHrdetectTsvFile$read()

UmHrdetectTsvFile$write()

UmHrdetectTsvFile$clone()

Inherited methods

dracarys::File$bname()

dracarys::File$initialize()

dracarys::File$print()

dracarys::File$type()

Method read()

Reads the hrdetect.tsv.gz file output from umccrise.

Usage

UmHrdetectTsvFile$read()

Returns

A tibble.

Method write()

Writes a tidy version of the hrdetect.tsv.gz file output from umccrise.

Usage

UmHrdetectTsvFile$write(d, out_dir, prefix, out_format = "tsv")

Arguments

d

Parsed object from self$read().

out_dir

Output directory.

prefix

Prefix of output file(s).

out_format

Format of output file(s) (one of 'tsv' (def.), 'parquet', 'both').

Method clone()

The objects of this class are cloneable with this method.

Usage

UmHrdetectTsvFile$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
x <- "/path/to/hrdetect.tsv.gz"
d <- UmHrdetectTsvFile$new(x)
d_parsed <- d$read() # or read(d)
d$write(d_parsed, out_dir = tempdir(), prefix = "sample705", out_format = "both")
}