VCMetricsFile R6 Class

Contains methods for reading and displaying contents of the vc_metrics.csv file output from DRAGEN, which contains variant calling metrics. reported for each sample in multi sample VCF and gVCF files. Based on the run case, metrics are reported either as standard VARIANT CALLER or JOINT CALLER. Metrics are reported both for the raw (PREFILTER) and hard filtered (POSTFILTER) VCFs. PON (Panel of Normals) and COSMIC filtered variants are counted as though they are PASS variants in the POSTFILTER VCFs metrics, which may result in higher than expected variant counts in the POSTFILTER VCF metrics.

Super class

dracarys::File -> VCMetricsFile

Methods

Public methods

VCMetricsFile$read()

VCMetricsFile$clone()

Inherited methods

dracarys::File$bname()

dracarys::File$initialize()

dracarys::File$print()

dracarys::File$type()

Method read()

Reads the vc_metrics.csv file output from DRAGEN.

Usage

VCMetricsFile$read()

Returns

tibble with the following columns:

category

sample

var: variable name

count: count value

pct: percent value

Method clone()

The objects of this class are cloneable with this method.

Usage

VCMetricsFile$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

x <- system.file("extdata/wgs/SEQC-II.vc_metrics.csv.gz", package = "dracarys")
vm <- VCMetricsFile$new(x)
vm$read() # or read(vm)
#> # A tibble: 45 × 5
#>    category  sample           var                    count    pct
#>    <chr>     <chr>            <chr>                  <dbl>  <dbl>
#>  1 summary   ""               Number of samples          1  NA   
#>  2 summary   ""               Reads Processed   2617162106  NA   
#>  3 summary   ""               Child Sample              NA  NA   
#>  4 prefilter "SEQC-II_tumour" Total                 177975 100   
#>  5 prefilter "SEQC-II_tumour" Biallelic             177975 100   
#>  6 prefilter "SEQC-II_tumour" Multiallelic               0   0   
#>  7 prefilter "SEQC-II_tumour" SNPs                  143829  80.8 
#>  8 prefilter "SEQC-II_tumour" Insertions (Hom)           0   0   
#>  9 prefilter "SEQC-II_tumour" Insertions (Het)       16354   9.19
#> 10 prefilter "SEQC-II_tumour" Deletions (Hom)            0   0   
#> # ℹ 35 more rows