bcl_convert Wf R6 Class

Contains methods for reading and processing files output from the UMCCR bcl_convert workflow.

Public fields

indir

Input directory containing Reports per assay type (e.g. /primary_data/240607_A01052_0209_BHLHFTDSXC/2024061140802544/).

Methods

Public methods

Wf_bcl_convert$new()

Wf_bcl_convert$print()

Wf_bcl_convert$clone()

Method new()

Create a new Wf_bcl_convert object.

Usage

Wf_bcl_convert$new(indir)

Arguments

indir

Input directory containing Reports per assay type.

Method print()

Print details about the Workflow

Usage

Wf_bcl_convert$print(...)

Arguments

...

(ignored).

Method clone()

The objects of this class are cloneable with this method.

Usage

Wf_bcl_convert$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) { # \dontrun{
indir <- file.path()
sample_id <- "PTC_ctTSO240429"
library_id <- "L2400482"
d <- TsoCombinedVariantOutputFile$new(x)
d$read()
} # }