sash Download Tidy and Write

Downloads files from the sash workflow and writes them in a tidy format.

Usage

Wf_sash_download_tidy_write(
  path,
  SubjectID,
  SampleID_tumor,
  outdir,
  format = "rds",
  max_files = 1000,
  ica_token = Sys.getenv("ICA_ACCESS_TOKEN"),
  dryrun = FALSE
)

Arguments

path

Path to directory with raw workflow results (from GDS, S3, or local filesystem).

SubjectID

The SubjectID of the sample (needed for path lookup).

SampleID_tumor

The SampleID of the tumor sample (needed for path lookup).

outdir

Path to output directory.

format

Format of output files.

max_files

Max number of files to list.

ica_token

ICA access token (def: $ICA_ACCESS_TOKEN env var).

dryrun

If TRUE, just list the files that will be downloaded (don't download them).

Value

List where each element is a tidy tibble of a sash file.

Examples

if (FALSE) { # \dontrun{
SubjectID <- "SBJ03043"
SampleID_tumor <- "PRJ230004"
p1_gds <- glue("gds://production/analysis_data/{SubjectID}/umccrise")
p <- file.path(p1_gds, "20240830ec648f40/L2300064__L2300063")
outdir <- file.path(sub("gds:/", "~/icav1/g", p))
token <- Sys.getenv("ICA_ACCESS_TOKEN")
d <- Wf_sash_download_tidy_write(
  path = p, SubjectID = SubjectID, SampleID_tumor = SampleID_tumor,
  outdir = outdir,
  dryrun = F
)
} # }