dracarys S3 Download

Download only S3 files that can be processed by dracarys.

Usage

dr_s3_download(
  s3dir,
  outdir,
  page_size = 100,
  pattern = NULL,
  regexes = DR_FILE_REGEX,
  dryrun = FALSE
)

Arguments

s3dir

Full path to S3 directory.

outdir

Path to output directory.

page_size

Page size (def: 100).

pattern

Pattern to further filter the returned file type tibble.

regexes

Tibble with regex and function name.

dryrun

If TRUE, just list the files that will be downloaded (don't download them).

Examples

s3dir <- file.path(
  "s3://umccr-primary-data-prod/UMCCR-Validation/SBJ00596",
  "ctTSO/2021-03-17/PTC_SSqCMM05pc_L2100067"
)
outdir <- sub("s3:/", "~/s3", s3dir)