List Files in ICAv1 GDS Directory

Lists files in a GDS directory.

Usage

gds_list_files_dir(
  gdsdir,
  token = Sys.getenv("ICA_ACCESS_TOKEN"),
  page_size = NULL,
  include_url = FALSE,
  no_recurse = TRUE,
  page_token = NULL,
  recursive = NULL
)

Arguments

gdsdir

Full path to GDS directory.

token

ICA access token (def: $ICA_ACCESS_TOKEN env var).

page_size

Page size (def: 10 via ICA API).

include_url

Include presigned URLs to all files within the GDS directory (def: FALSE via ICA API).

no_recurse

Do not recurse through the file list i.e. just give the first <page_size> items without recursing further down the list using <page_token>.

page_token

Page token (def: NULL). Used internally for recursion.

recursive

Should files be returned recursively in and under the specified GDS directory, or only directly in the specified GDS directory (def: TRUE via ICA API).

Value

A tibble with file ID, basename, size, last modified timestamp, full GDS path, and presigned URL if requested.

Examples

if (FALSE) { # \dontrun{
gdsdir <- file.path(
  "gds://production/analysis_data/SBJ00699/umccrise",
  "202203277dcf8562/L2200352__L2100146/SBJ00699__MDX220105/coverage"
)
token <- ica_token_validate()
page_size <- 11
include_url <- F
page_token <- NULL
no_recurse <- TRUE
recursive <- NULL
gds_list_files_dir(gdsdir, token, page_size, include_url, no_recurse, page_token, recursive)
} # }