MultiQC Extract Bar Graph Data

Extracts bar_graph data for the given plot name from a MultiQC JSON. Each bar_graph element in the JSON object contains:

Usage

multiqc_parse_bargraph_plot(dat)

Arguments

dat

Parsed JSON list element with specific plot data to extract.

Value

Tibble with name and data list-column.

Details

a samples array of N subarrays (of strings) of length A.

a datasets array of N subarrays (of objects) of length B. Each of the subarrays contains name (single string) and data (array of length equal to the length of the index-corresponding samples subarray).

Examples

if (FALSE) { # \dontrun{
j1 <- here::here("nogit/bcl_convert/multiqc_data.json")
j2 <- here::here("nogit/dracarys_gds_sync/multiqc_data.json")
j <- j1
j <- j2
multiqc_list_plots(j)
parsed <- RJSONIO::fromJSON(j)
dat <- parsed$report_plot_data$mapping_dup_percentage_plot
dat <- parsed$report_plot_data$time_metrics_plot
dat <- parsed$report_plot_data$bclconvert_lane_counts
dat <- parsed$report_plot_data$bclconvert_sample_counts
multiqc_parse_bargraph_plot(dat)
} # }