ReplayFile R6 Class

Contains methods for reading contents of the replay.json file output from DRAGEN, which contains the DRAGEN command line, parameters and version for the specific run.

Super class

dracarys::File -> ReplayFile

Methods

Public methods

ReplayFile$read()

ReplayFile$write()

ReplayFile$clone()

Inherited methods

dracarys::File$bname()

dracarys::File$initialize()

dracarys::File$print()

dracarys::File$type()

Method read()

Reads the replay.json file.

Usage

ReplayFile$read()

Returns

tibble with one row and metrics spread across individual columns.

Method write()

Writes a tidy version of the replay.json file output from DRAGEN.

Usage

ReplayFile$write(d, out_dir = NULL, prefix, out_format = "tsv", drid = NULL)

Arguments

d

Parsed object from self$read().

out_dir

Output directory.

prefix

Prefix of output file(s).

out_format

Format of output file(s).

drid

dracarys ID to use for the dataset (e.g. wfrid.123, prid.456).

Method clone()

The objects of this class are cloneable with this method.

Usage

ReplayFile$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

x <- system.file("extdata/wgs/SEQC-II-replay.json.gz", package = "dracarys")
r <- ReplayFile$new(x)
d <- r$read() # or read(r)
r$write(d, out_dir = tempdir(), prefix = "seqc_replay", out_format = "tsv")