TsoAlignCollapseFusionCallerMetricsFile R6 Class

Contains methods for reading and displaying contents of the AlignCollapseFusionCaller_metrics.json.gz file output from TSO.

Super class

dracarys::File -> TsoAlignCollapseFusionCallerMetricsFile

Methods

Public methods

TsoAlignCollapseFusionCallerMetricsFile$read()

TsoAlignCollapseFusionCallerMetricsFile$clone()

Inherited methods

dracarys::File$bname()

dracarys::File$initialize()

dracarys::File$print()

dracarys::File$type()

Method read()

Reads the AlignCollapseFusionCaller_metrics.json.gz file output from TSO.

Usage

TsoAlignCollapseFusionCallerMetricsFile$read()

Returns

tibble with the following columns:

section: name of original JSON element

name: name of metric

value: value of metric

percent: percentage

Method clone()

The objects of this class are cloneable with this method.

Usage

TsoAlignCollapseFusionCallerMetricsFile$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

x <- system.file("extdata/tso/sample705.AlignCollapseFusionCaller_metrics.json.gz",
  package = "dracarys"
)
m <- TsoAlignCollapseFusionCallerMetricsFile$new(x)
m$read() # or read(m)
#> # A tibble: 548 × 4
#>    section              name                                       value percent
#>    <chr>                <chr>                                      <chr> <chr>  
#>  1 MappingAligningPerRg Total reads in RG                          1063… 100    
#>  2 MappingAligningPerRg Number of duplicate marked reads           0     0      
#>  3 MappingAligningPerRg Number of duplicate marked and mate reads… 0     0      
#>  4 MappingAligningPerRg Number of unique reads (excl. duplicate m… 1063… 100    
#>  5 MappingAligningPerRg Reads with mate sequenced                  1063… 100    
#>  6 MappingAligningPerRg Reads without mate sequenced               0     0      
#>  7 MappingAligningPerRg QC-failed reads                            0     0      
#>  8 MappingAligningPerRg Mapped reads                               1057… 99.48  
#>  9 MappingAligningPerRg Mapped reads R1                            5294… 99.6   
#> 10 MappingAligningPerRg Mapped reads R2                            5281… 99.35  
#> # … with 538 more rows