TsoSampleAnalysisResultsFile R6 Class

Contains methods for reading and displaying contents of the SampleAnalysisResults.json.gz file output from TSO.

Super class

dracarys::File -> TsoSampleAnalysisResultsFile

Methods

Public methods

TsoSampleAnalysisResultsFile$read()

TsoSampleAnalysisResultsFile$write()

TsoSampleAnalysisResultsFile$clone()

Inherited methods

dracarys::File$bname()

dracarys::File$initialize()

dracarys::File$print()

dracarys::File$type()

Method read()

Reads the SampleAnalysisResults.json.gz file output from TSO.

Usage

TsoSampleAnalysisResultsFile$read()

Returns

list of tibbles

Method write()

Writes a tidy version of the SampleAnalysisResults.json.gz file output from TSO.

Usage

TsoSampleAnalysisResultsFile$write(d, out_dir, prefix, out_format = "tsv")

Arguments

d

Parsed object from self$read().

out_dir

Output directory.

prefix

Prefix of output file(s).

out_format

Format of output file(s) (one of 'tsv' (def.), 'parquet', 'both').

Method clone()

The objects of this class are cloneable with this method.

Usage

TsoSampleAnalysisResultsFile$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

x <- system.file("extdata/tso/sample705_SampleAnalysisResults.json.gz", package = "dracarys")
res <- TsoSampleAnalysisResultsFile$new(x)
d_parsed <- res$read() # or read(res)
res$write(d_parsed, tempfile(), "both")
#> $sample_info
#> # A tibble: 1 × 4
#>   sampleId                 analysisDate analysisTime analysisName              
#>   <chr>                    <chr>        <chr>        <chr>                     
#> 1 PTC_ctTSO220118_L2200054 2022-02-19   21:49:55     ctTSO220118-220207_18Feb22
#> 
#> $qc
#> # A tibble: 1 × 39
#>   dna_qc_CONTA…¹ dna_q…² dna_q…³ dna_q…⁴ dna_q…⁵ dna_q…⁶ dna_q…⁷ dna_q…⁸ dna_q…⁹
#>            <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1            664       0    1672       1       0   0.049    6131    1300      NA
#> # … with 30 more variables: dna_qc_smallv_tmb_PCT_EXON_1000X_value <dbl>,
#> #   dna_qc_smallv_tmb_PCT_EXON_1000X_LSL <dbl>,
#> #   dna_qc_smallv_tmb_PCT_EXON_1000X_USL <dbl>,
#> #   dna_qc_cnv_GENE_SCALED_MAD_value <dbl>,
#> #   dna_qc_cnv_GENE_SCALED_MAD_LSL <dbl>, dna_qc_cnv_GENE_SCALED_MAD_USL <dbl>,
#> #   dna_qc_cnv_MEDIAN_BIN_COUNT_CNV_TARGET_value <dbl>,
#> #   dna_qc_cnv_MEDIAN_BIN_COUNT_CNV_TARGET_LSL <dbl>, …
#> 
#> $biomarkers
#> # A tibble: 1 × 5
#>   msi_pct_unstable_sites msi_SumJsd tmb_per_mb tmb_coding_region_sizemb tmb_so…¹
#>                    <dbl>      <dbl>      <dbl>                    <dbl>    <dbl>
#> 1                    2.2       1.40       144.                     1.22      176
#> # … with abbreviated variable name ¹​tmb_somatic_coding_variants_count
#> 
#> $sw_conf_datasources
#> # A tibble: 7 × 4
#>   name                 version     description                           relea…¹
#>   <chr>                <chr>       <chr>                                 <chr>  
#> 1 VEP                  91          BothRefSeqAndEnsembl                  2018-0…
#> 2 1000 Genomes Project Phase 3 v5a A public catalogue of human variatio… 2013-0…
#> 3 ClinVar              20190204    A freely accessible, public archive … 2019-0…
#> 4 COSMIC               84          somatic mutation and related details… 2018-0…
#> 5 dbSNP                151         Identifiers for observed variants     2018-0…
#> 6 gnomAD               2.1         Whole genome allele frequencies from… 2018-1…
#> 7 gnomAD_exome         2.1         Exome allele frequencies from Genome… 2018-1…
#> # … with abbreviated variable name ¹​releaseDate
#> 
#> $sw_conf_other
#> # A tibble: 1 × 7
#>   analysisSoftwareName analysisSoftwar…¹ genom…² insta…³ nirva…⁴ genom…⁵ refse…⁶
#>   <chr>                <chr>             <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 TSO500_ctDNA_RnD     0.0.0.1           hg19_d… CDxNir… 3.2.3   GRCh37  2015-01
#> # … with abbreviated variable names ¹​analysisSoftwareVersion, ²​genomeBuild,
#> #   ³​instanceName, ⁴​nirvanaSoftwareVersion, ⁵​genomeAssembly, ⁶​refseqVersion
#> 
#> $snv
#> # A tibble: 1,980 × 29
#>    chrom      pos ref   alt       af  qual dp_tot dp_alt transc…¹ source bioType
#>    <chr>    <int> <chr> <chr>  <dbl> <dbl>  <int>  <int> <chr>    <chr>  <chr>  
#>  1 chr1   2488153 A     G     0.488    100   4917   2402 NM_0038… RefSeq protei…
#>  2 chr1   4367323 G     A     1.00     100   5272   5271 NA       NA     NA     
#>  3 chr1   9778888 G     A     0.0016   100   7923     13 NM_0050… RefSeq protei…
#>  4 chr1   9781605 G     A     0.0017   100   7471     13 NM_0050… RefSeq protei…
#>  5 chr1   9781637 G     A     0.0013   100   6136      8 NM_0050… RefSeq protei…
#>  6 chr1   9784354 C     T     0.0019   100   4699      9 NM_0050… RefSeq protei…
#>  7 chr1  11205058 C     T     1.00     100   5856   5854 NM_0049… RefSeq protei…
#>  8 chr1  11276255 C     G     0.0032   100    950      3 NM_0049… RefSeq protei…
#>  9 chr1  11288758 G     A     1.00     100   8997   8995 NM_0049… RefSeq protei…
#> 10 chr1  11292504 C     T     0.0021   100   3809      8 NM_0049… RefSeq protei…
#> # … with 1,970 more rows, 18 more variables: aminoAcids <chr>, cdnaPos <chr>,
#> #   codons <chr>, cdsPos <chr>, exons <chr>, geneId <chr>, hgnc <chr>,
#> #   hgvsc <chr>, hgvsp <chr>, isCanonical <lgl>, polyPhenScore <dbl>,
#> #   polyPhenPrediction <chr>, proteinId <chr>, proteinPos <chr>,
#> #   siftScore <dbl>, siftPrediction <chr>, consequence <chr>, introns <chr>,
#> #   and abbreviated variable name ¹​transcript
#> 
#> $cnv
#> # A tibble: 1 × 7
#>   foldChange  qual copyNumberType gene  chromosome startPosition endPosition
#>        <dbl> <dbl> <chr>          <chr> <chr>              <int>       <int>
#> 1       0.86    57 <DEL>          ATM   chr11          108098348   108236237
#>