TsoSampleAnalysisResultsFile R6 Class

Contains methods for reading and displaying contents of the SampleAnalysisResults.json.gz file output from TSO.

Super class

dracarys::File -> TsoSampleAnalysisResultsFile

Methods

Public methods

TsoSampleAnalysisResultsFile$read()

TsoSampleAnalysisResultsFile$clone()

Inherited methods

dracarys::File$bname()

dracarys::File$initialize()

dracarys::File$print()

dracarys::File$type()

Method read()

Reads the SampleAnalysisResults.json.gz file output from TSO.

Usage

TsoSampleAnalysisResultsFile$read()

Returns

list of tibbles

Method clone()

The objects of this class are cloneable with this method.

Usage

TsoSampleAnalysisResultsFile$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

x <- system.file("extdata/tso/sample705_SampleAnalysisResults.json.gz", package = "dracarys")
res <- TsoSampleAnalysisResultsFile$new(x)
res$read() # or read(res)
#> $sample_info
#> # A tibble: 1 × 4
#>   sampleId                 analysisDate analysisTime analysisName              
#>   <chr>                    <chr>        <chr>        <chr>                     
#> 1 PTC_ctTSO220118_L2200054 2022-02-19   21:49:55     ctTSO220118-220207_18Feb22
#> 
#> $software_config
#> $software_config$data_sources
#> # A tibble: 7 × 4
#>   name                 version     description                           relea…¹
#>   <chr>                <chr>       <chr>                                 <chr>  
#> 1 VEP                  91          BothRefSeqAndEnsembl                  2018-0…
#> 2 1000 Genomes Project Phase 3 v5a A public catalogue of human variatio… 2013-0…
#> 3 ClinVar              20190204    A freely accessible, public archive … 2019-0…
#> 4 COSMIC               84          somatic mutation and related details… 2018-0…
#> 5 dbSNP                151         Identifiers for observed variants     2018-0…
#> 6 gnomAD               2.1         Whole genome allele frequencies from… 2018-1…
#> 7 gnomAD_exome         2.1         Exome allele frequencies from Genome… 2018-1…
#> # … with abbreviated variable name ¹​releaseDate
#> 
#> $software_config$other
#> # A tibble: 1 × 7
#>   analysisSoftwareName analysisSoftwar…¹ genom…² insta…³ nirva…⁴ genom…⁵ refse…⁶
#>   <chr>                <chr>             <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 TSO500_ctDNA_RnD     0.0.0.1           hg19_d… CDxNir… 3.2.3   GRCh37  2015-01
#> # … with abbreviated variable names ¹​analysisSoftwareVersion, ²​genomeBuild,
#> #   ³​instanceName, ⁴​nirvanaSoftwareVersion, ⁵​genomeAssembly, ⁶​refseqVersion
#> 
#> 
#> $biomarkers
#> $biomarkers$msi
#> $biomarkers$msi$msi_pct_unstable_sites
#> [1] 2.2
#> 
#> $biomarkers$msi$additional_metrics
#> # A tibble: 1 × 3
#>   name   value   RAD
#>   <chr>  <dbl> <dbl>
#> 1 SumJsd  1.40     3
#> 
#> 
#> $biomarkers$tmb
#> $biomarkers$tmb$tmb_per_mb
#> [1] 143.8
#> 
#> $biomarkers$tmb$additional_metrics
#> # A tibble: 2 × 3
#>   name                        value   RAD
#>   <chr>                       <dbl> <dbl>
#> 1 CodingRegionSizeMb           1.22     2
#> 2 SomaticCodingVariantsCount 176        0
#> 
#> 
#> 
#> $sample_metrics_qc
#> # A tibble: 7 × 7
#>   metric                                name    value   LSL      USL UOM     RAD
#>   <chr>                                 <chr>   <dbl> <dbl>    <dbl> <chr> <dbl>
#> 1 DNA Library QC Metrics                CONT… 6.64e+2     0 1672     NA        0
#> 2 DNA Library QC Metrics                CONT… 1   e+0     0    0.049 NA        3
#> 3 DNA Library QC Metrics for Small Var… MEDI… 6.13e+3  1300   NA     count     0
#> 4 DNA Library QC Metrics for Small Var… PCT_… 9.73e+1    80   NA     %         1
#> 5 DNA Library QC Metrics for Copy Numb… GENE… 8.8 e-2     0    0.059 count     3
#> 6 DNA Library QC Metrics for Copy Numb… MEDI… 5.31e+1     6   NA     count     1
#> 7 DNA Library QC Metrics for MSI and F… MEDI… 6.13e+3  1300   NA     count     0
#> 
#> $sample_metrics_expanded
#> # A tibble: 1 × 18
#>   TOTAL_PF_READS MEAN_…¹ MEDIA…² PCT_C…³ PCT_E…⁴ PCT_E…⁵ PCT_R…⁶ PCT_U…⁷ MEAN_…⁸
#>            <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1     1063188966     4.1    5845    0.06    98.4    95.5    81.1    99.9    5937
#> # … with 9 more variables: PCT_ALIGNED_READS <dbl>,
#> #   PCT_CONTAMINATION_EST <dbl>, PCT_TARGET_0.4X_MEAN <dbl>,
#> #   PCT_TARGET_500X <dbl>, PCT_TARGET_1000X <dbl>, PCT_TARGET_1500X <dbl>,
#> #   PCT_DUPLEXFAMILIES <dbl>, MEDIAN_INSERT_SIZE <dbl>, MAX_SOMATIC_AF <dbl>,
#> #   and abbreviated variable names ¹​MEAN_FAMILY_SIZE, ²​MEDIAN_TARGET_COVERAGE,
#> #   ³​PCT_CHIMERIC_READS, ⁴​PCT_EXON_500X, ⁵​PCT_EXON_1500X, ⁶​PCT_READ_ENRICHMENT,
#> #   ⁷​PCT_USABLE_UMI_READS, ⁸​MEAN_TARGET_COVERAGE
#> 
#> $vars
#> $vars$snvs
#> # A tibble: 1,980 × 29
#>    chrom      pos ref   alt       af  qual dp_tot dp_alt transc…¹ source bioType
#>    <chr>    <int> <chr> <chr>  <dbl> <dbl>  <int>  <int> <chr>    <chr>  <chr>  
#>  1 chr1   2488153 A     G     0.488    100   4917   2402 NM_0038… RefSeq protei…
#>  2 chr1   4367323 G     A     1.00     100   5272   5271 NA       NA     NA     
#>  3 chr1   9778888 G     A     0.0016   100   7923     13 NM_0050… RefSeq protei…
#>  4 chr1   9781605 G     A     0.0017   100   7471     13 NM_0050… RefSeq protei…
#>  5 chr1   9781637 G     A     0.0013   100   6136      8 NM_0050… RefSeq protei…
#>  6 chr1   9784354 C     T     0.0019   100   4699      9 NM_0050… RefSeq protei…
#>  7 chr1  11205058 C     T     1.00     100   5856   5854 NM_0049… RefSeq protei…
#>  8 chr1  11276255 C     G     0.0032   100    950      3 NM_0049… RefSeq protei…
#>  9 chr1  11288758 G     A     1.00     100   8997   8995 NM_0049… RefSeq protei…
#> 10 chr1  11292504 C     T     0.0021   100   3809      8 NM_0049… RefSeq protei…
#> # … with 1,970 more rows, 18 more variables: aminoAcids <chr>, cdnaPos <chr>,
#> #   codons <chr>, cdsPos <chr>, exons <chr>, geneId <chr>, hgnc <chr>,
#> #   hgvsc <chr>, hgvsp <chr>, isCanonical <lgl>, polyPhenScore <dbl>,
#> #   polyPhenPrediction <chr>, proteinId <chr>, proteinPos <chr>,
#> #   siftScore <dbl>, siftPrediction <chr>, consequence <chr>, introns <chr>,
#> #   and abbreviated variable name ¹​transcript
#> 
#> $vars$cnvs
#> # A tibble: 1 × 7
#>   foldChange  qual copyNumberType gene  chromosome startPosition endPosition
#>        <dbl> <dbl> <chr>          <chr> <chr>              <int>       <int>
#> 1       0.86    57 <DEL>          ATM   chr11          108098348   108236237
#> 
#>