TsoTargetRegionCoverageFile R6 Class

Contains methods for reading and displaying contents of the TargetRegionCoverage.json.gz file output from TSO.

Super class

dracarys::File -> TsoTargetRegionCoverageFile

Methods

Public methods

TsoTargetRegionCoverageFile$read()

TsoTargetRegionCoverageFile$write()

TsoTargetRegionCoverageFile$plot()

TsoTargetRegionCoverageFile$clone()

Inherited methods

dracarys::File$bname()

dracarys::File$initialize()

dracarys::File$print()

dracarys::File$type()

Method read()

Reads the TargetRegionCoverage.json.gz file output from TSO.

Usage

TsoTargetRegionCoverageFile$read()

Returns

tibble with the following columns:

ConsensusReadDepth

BasePair

Percentage

Method write()

Writes a tidy version of the TargetRegionCoverage.json.gz file output from TSO.

Usage

TsoTargetRegionCoverageFile$write(d, out_dir, prefix, out_format = "tsv")

Arguments

d

Parsed object from self$read().

out_dir

Output directory.

prefix

Prefix of output file(s).

out_format

Format of output file(s) (one of 'tsv' (def.), 'parquet', 'both').

Method plot()

Plots the TargetRegionCoverage.json.gz file.

Usage

TsoTargetRegionCoverageFile$plot(d, min_pct = 2)

Arguments

d

Parsed object from self$read().

min_pct

Minimum percentage to be plotted (Default: 2).

Returns

A ggplot2 plot containing read depth on X axis and percentage covered on Y axis.

Method clone()

The objects of this class are cloneable with this method.

Usage

TsoTargetRegionCoverageFile$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

x <- system.file("extdata/tso/sample705.TargetRegionCoverage.json.gz", package = "dracarys")
trc <- TsoTargetRegionCoverageFile$new(x)
d_parsed <- trc$read() # or read(trc)
trc$plot(d_parsed, 90) # or plot(trc, d_parsed, 90)
#> Error in loadNamespace(x): there is no package called ‘ggrepel’
trc$write(d_parsed, out_dir = tempdir(), prefix = "sample705", out_format = "both")
#> # A tibble: 14 × 3
#>    ConsensusReadDepth BasePair Percentage
#>    <chr>              <chr>         <dbl>
#>  1 TargetRegion       1970825      100   
#>  2 100X               1953719       99.1 
#>  3 250X               1940233       98.4 
#>  4 500X               1922662       97.6 
#>  5 750X               1907303       96.8 
#>  6 1000X              1891110       96.0 
#>  7 1500X              1848614       93.8 
#>  8 2000X              1795480       91.1 
#>  9 2500X              1729811       87.8 
#> 10 3000X              1651245       83.8 
#> 11 4000X              1445949       73.4 
#> 12 5000X              1209504       61.4 
#> 13 8000X              522143        26.5 
#> 14 10000X             192278         9.76