Workflow R6 Class

Workflow is a base R6 class representing a bioinformatic workflow run from a UMCCR workflow manager.

Public fields

prid

Portal run ID.

type

Workflow type.

start

Workflow start datetime.

end

Workflow end datetime.

status

Workflow end status.

input

Workflow input JSON string.

output

Workflow output JSON string.

Methods

Public methods

Wf$new()

Wf$print()

Wf$clone()

Method new()

Create a new Workflow object.

Usage

Wf$new(
  prid = NULL,
  type = NULL,
  start = NULL,
  end = NULL,
  status = NULL,
  input = NULL,
  output = NULL
)

Arguments

prid

Portal run ID.

type

Workflow type.

start

Workflow start datetime.

end

Workflow end datetime.

status

Workflow end status.

input

Workflow input JSON string.

output

Workflow output JSON string.

Method print()

Print details about the Workflow.

Usage

Wf$print(...)

Arguments

...

(ignored).

Method clone()

The objects of this class are cloneable with this method.

Usage

Wf$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

p1 <- system.file("extdata/portaldb_workflow_top4.rds", package = "rportal") |>
  readRDS() |>
  dplyr::filter(type_name == "umccrise") |>
  dplyr::slice(1)
#> Warning: cannot open compressed file '', probable reason 'No such file or directory'
#> Error in gzfile(file, "rb"): cannot open the connection
w <- Wf$new(
  prid = p1$portal_run_id, type = p1$type_name, start = p1$start, end = p1$end,
  status = p1$end_status, input = p1$input, output = p1$output
)
#> Error in eval(expr, envir, enclos): object 'p1' not found
w
#> Error in eval(expr, envir, enclos): object 'w' not found